pyproject.toml
[project]
name = "mechgaia-agentbeats"
version = "0.1.0"
description = "MechGaia agent assessment framework using A2A and MCP standards."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.8",
    "dotenv>=0.9.9",
    "earthshaker>=0.1.12",
    "pydantic-settings>=2.11.0",
    "typer>=0.19.2",
    "uvicorn>=0.37.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "litellm>=1.60.1",
    "sqlalchemy>=2.0.0",
    "sympy>=1.12",
    "matplotlib>=3.8.0",
    "scikit-learn>=1.3.0",
]
